Consensus finally works
While hashing the block inside the ProofOfAuthority() function, the
block was being hashed before signing the block. But VerifyProof()
function verifies the hash after the block has a signature. So we set
the signature field empty while hashing the block. This commit message
could be placed on block.go but fine.